Inferential Statistics for VCA-Results

A copy from VCA::VCAinference in VCA package

Usage

VCAinference(...)

Arguments

...

Arguments passed on to VCA::VCAinference

obj

(object) of class 'VCA' or, alternatively, a list of 'VCA' objects, where all other arguments can be specified as vectors, where the i-th vector element applies to the i-th element of 'obj' (see examples)

alpha

(numeric) value specifying the significance level for \(100*(1-alpha)\)% confidence intervals.

total.claim

(numeric) value specifying the claim-value for the Chi-Squared test for the total variance (SD or CV, see claim.type).

error.claim

(numeric) value specifying the claim-value for the Chi-Squared test for the error variance (SD or CV, see claim.type).

claim.type

(character) one of "VC", "SD", "CV" specifying how claim-values have to be interpreted:
"VC" (Default) = claim-value(s) specified in terms of variance(s),
"SD" = claim-values specified in terms of standard deviations (SD),
"CV" = claim-values specified in terms of coefficient(s) of variation (CV) and are specified as percentages.
If set to "SD" or "CV", claim-values will be converted to variances before applying the Chi-Squared test (see examples).

VarVC

(logical) TRUE = the covariance matrix of the estimated VCs will be computed (see vcovVC), where diagonal elements correspond to the variances of the individual VCs. This matrix is required for estimation of CIs for intermediate VCs if 'method.ci="sas"'. FALSE (Default) = computing covariance matrix of VCs is omitted, as well as CIs for intermediate VCs.

excludeNeg

(logical) TRUE = confidence intervals of negative variance estimates will not be reported.
FALSE = confidence intervals for all VCs will be reported including those with negative VCs.
See the details section for a thorough explanation.

constrainCI

(logical) TRUE = CI-limits for all variance components are constrained to be >= 0.
FALSE = unconstrained CIs with potentially negative CI-limits will be reported.
which will preserve the original width of CIs. See the details section for a thorough explanation.

ci.method

(character) string or abbreviation specifying which approach to use for computing confidence intervals of variance components (VC). "sas" (default) uses Chi-Squared based CIs for total and error and normal approximation for all other VCs (Wald-limits, option "NOBOUND" in SAS PROC MIXED); "satterthwaite" will approximate DFs for each VC using the Satterthwaite approach (see SattDF for models fitted by ANOVA) and all Cis are based on the Chi-Squared distribution. This approach is conservative but avoids negative values for the lower bounds.

quiet

(logical) TRUE = will suppress any warning, which will be issued otherwise

Value

object of VCAinference contains a series of statistics.

See also

Examples

data(glucose)
fit <- anovaVCA(value ~ day / run, glucose)
VCAinference(fit)
#> 
#> 
#> 
#> Inference from (V)ariance (C)omponent (A)nalysis
#> ------------------------------------------------
#> 
#> > VCA Result:
#> -------------
#> 
#>   Name    DF      SS    MS      VC      %Total  SD     CV[%] 
#> 1 total   64.7773               12.9336 100     3.5963 1.4727
#> 2 day     19      415.8 21.8842 1.9586  15.1432 1.3995 0.5731
#> 3 day:run 20      281   14.05   3.075   23.7754 1.7536 0.7181
#> 4 error   40      316   7.9     7.9     61.0814 2.8107 1.151 
#> 
#> Mean: 244.2 (N = 80) 
#> 
#> Experimental Design: balanced  |  Method: ANOVA
#> 
#> 
#> > VC:
#> -----
#>         Estimate CI LCL  CI UCL One-Sided LCL One-Sided UCL
#> total    12.9336 9.4224 18.8614        9.9071       17.7278
#> day       1.9586                                           
#> day:run   3.0750                                           
#> error     7.9000 5.3251 12.9333        5.6673       11.9203
#> 
#> > SD:
#> -----
#>         Estimate CI LCL CI UCL One-Sided LCL One-Sided UCL
#> total     3.5963 3.0696 4.3430        3.1476        4.2104
#> day       1.3995                                          
#> day:run   1.7536                                          
#> error     2.8107 2.3076 3.5963        2.3806        3.4526
#> 
#> > CV[%]:
#> --------
#>         Estimate CI LCL CI UCL One-Sided LCL One-Sided UCL
#> total     1.4727  1.257 1.7785        1.2889        1.7242
#> day       0.5731                                          
#> day:run   0.7181                                          
#> error     1.1510  0.945 1.4727        0.9749        1.4138
#> 
#> 
#> 95% Confidence Level  
#> SAS PROC MIXED method used for computing CIs 
#>