ANOVA-Type Estimation of Variance Components for Random Models

A copy from VCA::anovaVCA in VCA package

Usage

anovaVCA(...)

Arguments

...

Arguments passed on to VCA::anovaVCA

form

(formula) specifying the model to be fit, a response variable left of the '~' is mandatory

Data

(data.frame) containing all variables referenced in 'form'

by

(factor, character) variable specifying groups for which the analysis should be performed individually, i.e. by-processing

NegVC

(logical) FALSE = negative variance component estimates (VC) will be set to 0 and they will not contribute to the total variance (as done in SAS PROC NESTED, conservative estimate of total variance). The original ANOVA estimates can be found in element 'VCoriginal'. The degrees of freedom of the total variance are based on adapted mean squares (MS), i.e. adapted MS are computed as \(D * VC\), where VC is the column vector with negative VCs set to 0.
TRUE = negative variance component estimates will not be set to 0 and they will contribute to the total variance (original definition of the total variance).

VarVC.method

(character) string specifying whether to use the algorithm given in Searle et al. (1992) which corresponds to VarVC.method="scm" or in Giesbrecht and Burns (1985) which can be specified via "gb". Method "scm" (Searle, Casella, McCulloch) is the exact algorithm, "gb" (Giesbrecht, Burns) is termed "rough approximation" by the authors, but sufficiently exact compared to e.g. SAS PROC MIXED (method=type1) which uses the inverse of the Fisher-Information matrix as approximation. For balanced designs all methods give identical results, only in unbalanced designs differences occur.

MME

(logical) TRUE = (M)ixed (M)odel (E)quations will be solved, i.e. 'VCA' object will have additional elements "RandomEffects", "FixedEffects", "VarFixed" (variance-covariance matrix of fixed effects) and the "Matrices" element has addional elements corresponding to intermediate results of solving MMEs. FALSE = do not solve MMEs, which reduces the computation time for very complex models significantly.

quiet

(logical) TRUE = will suppress any warning, which will be issued otherwise

order.data

(logical) TRUE = class-variables will be ordered increasingly, FALSE = ordering of class-variables will remain as is

Value

a class of VCA for downstream analysis.

See also

Examples

data(glucose)
anovaVCA(value ~ day / run, glucose)
#> 
#> 
#> Result Variance Component Analysis:
#> -----------------------------------
#> 
#>   Name    DF       SS    MS        VC        %Total    SD       CV[%]   
#> 1 total   64.77732                 12.933553 100       3.596325 1.472697
#> 2 day     19       415.8 21.884211 1.958553  15.143191 1.399483 0.573089
#> 3 day:run 20       281   14.05     3.075     23.77537  1.753568 0.718087
#> 4 error   40       316   7.9       7.9       61.081439 2.810694 1.15098 
#> 
#> Mean: 244.2 (N = 80) 
#> 
#> Experimental Design: balanced  |  Method: ANOVA
#>