Systematical Bias Between Reference Method and Test Method

A copy from mcr::calcBias in mcr package

Usage

calcBias(...)

Arguments

...

Arguments passed on to mcr::calcBias

.Object

object of class "MCResult".

Value

Bis and corresponding confidence interval for the specific medical decision levels (x.levels).

See also

Examples

data(platelet)
fit <- mcreg(
  x = platelet$Comparative, y = platelet$Candidate,
  method.reg = "Deming", method.ci = "jackknife"
)
#> Jackknife based calculation of standard error and confidence intervals according to Linnet's method.
calcBias(fit, x.levels = c(30, 200))
#>    Level     Bias       SE      LCI      UCI
#> X1    30 4.724429 1.378232 1.995155 7.453704
#> X2   200 6.926183 1.288534 4.374535 9.477832
calcBias(fit, x.levels = c(30, 200), type = "proportional")
#>    Level Prop.bias(%)        SE      LCI       UCI
#> X1    30    15.748098 4.5941062 6.650517 24.845679
#> X2   200     3.463092 0.6442671 2.187267  4.738916
calcBias(fit, x.levels = c(30, 200), type = "proportional", percent = FALSE)
#>    Level  Prop.bias          SE        LCI        UCI
#> X1    30 0.15748098 0.045941062 0.06650517 0.24845679
#> X2   200 0.03463092 0.006442671 0.02187267 0.04738916