Creates Contingency Table

Creates a 2x2 contingency table from the data frame or matrix for the qualitative performance of downstream analysis.

Usage

diagTab(formula = ~., data, ..., rlevels = NULL, clevels = NULL)

Arguments

formula

(numeric)
a formula object with the cross-classifying variables (separated by +) on the right hand side. The row name of contingency is represented by the variable to the left of the + sign, and the col name is represented by the variable to the right.

data

(data.frame or matrix)
a data frame or matrix.

...

other arguments to be passed to xtabs().

rlevels

(vector)
a character vector of known levels for row name of contingency table.

clevels

(vector)
a character vector of known levels for col name of contingency table.

Value

A object matrix contains the 2x2 contingency table.

See also

Examples

data("qualData")
qualData %>% diagTab(formula = ~ CandidateN + ComparativeN)
#>  Contingency Table: 
#> 
#>   candidate has 2 levels: 0 1
#>   comparative has 2 levels: 0 1 
#> 
#>           ComparativeN
#> CandidateN  0   1
#>          0 54  16
#>          1  8 122
qualData %>%
  diagTab(
    formula = ~ CandidateN + ComparativeN,
    rlevels = c(1, 0), clevels = c(1, 0)
  )
#>  Contingency Table: 
#> 
#>   candidate has 2 levels: 1 0
#>   comparative has 2 levels: 1 0 
#> 
#>           ComparativeN
#> CandidateN   1  0
#>          1 122  8
#>          0  16 54